Add Modbus Master plugin types and utility functions
- Introduced `modbus_master_types.py` with data classes for Modbus connection configuration, I/O point representation, device configuration, and buffer access details.
- Created `modbus_master_utils.py` containing utility functions for GCD calculation, batch read/write request generation, IEC size handling, and conversion between Modbus registers and IEC values.
- Updated `modbus_master_config_model.py` to integrate cycle time configuration into I/O points and validate cycle times during device configuration.
