+++ core/src/drivers/plugins/python/modbus_master/modbus_master_plugin.py
@@ -6,179 +6,19 @@ import re
 import threading
 import time
 import math
+import traceback
 from typing import Optional, Literal, List, Dict, Any
 
 from pymodbus.client import ModbusTcpClient
 from pymodbus.exceptions import ModbusIOException, ConnectionException
 from pymodbus.pdu import ExceptionResponse
 
-def gcd(a: int, b: int) -> int:
-    """
-    Calculate the Greatest Common Divisor of two numbers using Euclidean algorithm.
-    """
-    while b != 0:
-        a, b = b, a % b
-    return a
-
-def calculate_gcd_of_cycle_times(io_points: List[Any]) -> int:
-    """
-    Calculate the GCD of all cycle_time_ms values from I/O points.
-    If no points have cycle_time_ms, return 1000 (1 second default).
-    """
-    cycle_times = []
-    for point in io_points:
-        if hasattr(point, 'cycle_time_ms') and point.cycle_time_ms > 0:
-            cycle_times.append(point.cycle_time_ms)
-
-    if not cycle_times:
-        return 1000  # Default 1 second
-
-    # Calculate GCD of all cycle times
-    result = cycle_times[0]
-    for time_ms in cycle_times[1:]:
-        result = gcd(result, time_ms)
-
-    return result
-
-def get_batch_read_requests_from_io_points(io_points: List[Any]) -> Dict[int, List[Any]]:
-    """
-    Groups I/O points by Modbus read function code (1,2,3,4) and creates
-    batch read lists to optimize Modbus operations.
-    Returns a dictionary mapping FC to lists of points.
-    """
-    read_requests: Dict[int, List[Any]] = {}
-    for point in io_points:
-        fc = point.fc
-        if fc in [1, 2, 3, 4]:  # Read functions
-            if fc not in read_requests:
-                read_requests[fc] = []
-            read_requests[fc].append(point)
-    return read_requests
-
-def get_batch_write_requests_from_io_points(io_points: List[Any]) -> Dict[int, List[Any]]:
-    """
-    Groups I/O points by Modbus write function code (5,6,15,16) and creates
-    batch write lists to optimize Modbus operations.
-    Returns a dictionary mapping FC to lists of points.
-    """
-    write_requests: Dict[int, List[Any]] = {}
-    for point in io_points:
-        fc = point.fc
-        if fc in [5, 6, 15, 16]:  # Write functions
-            if fc not in write_requests:
-                write_requests[fc] = []
-            write_requests[fc].append(point)
-    return write_requests
-
-def get_modbus_registers_count_for_iec_size(iec_size: str) -> int:
-    """
-    Returns how many 16-bit Modbus registers are needed for an IEC data type.
-    
-    Args:
-        iec_size: IEC data size ('X', 'B', 'W', 'D', 'L')
-    
-    Returns:
-        Number of 16-bit registers needed
-    """
-    if iec_size == "X":  # 1 bit - handled separately (coils/discrete inputs)
-        return 0  # Not applicable for registers
-    elif iec_size == "B":  # 8 bits - fits in 1 register (with some unused bits)
-        return 1
-    elif iec_size == "W":  # 16 bits - exactly 1 register
-        return 1
-    elif iec_size == "D":  # 32 bits - needs 2 registers
-        return 2
-    elif iec_size == "L":  # 64 bits - needs 4 registers
-        return 4
-    else:
-        return 1  # Default fallback
-
-def convert_modbus_registers_to_iec_value(registers: List[int], iec_size: str, use_big_endian: bool = False):
-    """
-    Converts Modbus register values to IEC data type value.
-    
-    Args:
-        registers: List of 16-bit register values from Modbus
-        iec_size: IEC data size ('B', 'W', 'D', 'L')
-        use_big_endian: If True, use big-endian byte order, else little-endian
-    
-    Returns:
-        Converted value ready for IEC buffer
-    """
-    if iec_size == "B":  # 8 bits
-        # Take lower 8 bits of first register
-        return registers[0] & 0xFF
-    elif iec_size == "W":  # 16 bits
-        # Single register, no conversion needed
-        return registers[0] & 0xFFFF
-    elif iec_size == "D":  # 32 bits
-        # Combine 2 registers into 32-bit value
-        if len(registers) < 2:
-            raise ValueError("Need at least 2 registers for D (32-bit) type")
-        if use_big_endian:
-            return (registers[0] << 16) | registers[1]
-        else:  # little-endian
-            return (registers[1] << 16) | registers[0]
-    elif iec_size == "L":  # 64 bits
-        # Combine 4 registers into 64-bit value
-        if len(registers) < 4:
-            raise ValueError("Need at least 4 registers for L (64-bit) type")
-        if use_big_endian:
-            return (registers[0] << 48) | (registers[1] << 32) | (registers[2] << 16) | registers[3]
-        else:  # little-endian
-            return (registers[3] << 48) | (registers[2] << 32) | (registers[1] << 16) | registers[0]
-    else:
-        raise ValueError(f"Unsupported IEC size for register conversion: {iec_size}")
-
-def convert_iec_value_to_modbus_registers(value: int, iec_size: str, use_big_endian: bool = False) -> List[int]:
-    """
-    Converts IEC data type value to Modbus register values.
-    
-    Args:
-        value: IEC value to convert
-        iec_size: IEC data size ('B', 'W', 'D', 'L')
-        use_big_endian: If True, use big-endian byte order, else little-endian
-    
-    Returns:
-        List of 16-bit register values for Modbus
-    """
-    if iec_size == "B":  # 8 bits
-        # Put 8-bit value in lower part of register, upper part is 0
-        return [value & 0xFF]
-    elif iec_size == "W":  # 16 bits
-        # Single register
-        return [value & 0xFFFF]
-    elif iec_size == "D":  # 32 bits
-        # Split into 2 registers
-        if use_big_endian:
-            return [(value >> 16) & 0xFFFF, value & 0xFFFF]
-        else:  # little-endian
-            return [value & 0xFFFF, (value >> 16) & 0xFFFF]
-    elif iec_size == "L":  # 64 bits
-        # Split into 4 registers
-        if use_big_endian:
-            return [
-                (value >> 48) & 0xFFFF,
-                (value >> 32) & 0xFFFF,
-                (value >> 16) & 0xFFFF,
-                value & 0xFFFF
-            ]
-        else:  # little-endian
-            return [
-                value & 0xFFFF,
-                (value >> 16) & 0xFFFF,
-                (value >> 32) & 0xFFFF,
-                (value >> 48) & 0xFFFF
-            ]
-    else:
-        raise ValueError(f"Unsupported IEC size for register conversion: {iec_size}")
-
 # Add the parent directory to Python path to find shared module
 sys.path.insert(0, os.path.join(os.path.dirname(__file__), '..'))
 
 # Import the correct type definitions
 from shared.python_plugin_types import (
-    PluginRuntimeArgs, 
+    PluginRuntimeArgs,
     safe_extract_runtime_args_from_capsule,
     SafeBufferAccess,
     PluginStructureValidator
@@ -186,6 +26,38 @@ from shared.python_plugin_types import (
 
 # Import the configuration model
 from shared.plugin_config_decode.modbus_master_config_model import ModbusMasterConfig
+
+# Import local modules
+try:
+    # Try relative imports first (when used as package)
+    from .modbus_master_types import ModbusDeviceConfig
+    from .modbus_master_utils import (
+        calculate_gcd_of_cycle_times,
+        get_batch_read_requests_from_io_points,
+        get_batch_write_requests_from_io_points,
+        get_modbus_registers_count_for_iec_size,
+        parse_modbus_offset
+    )
+    from .modbus_master_memory import (
+        update_iec_buffer_from_modbus_data,
+        read_data_for_modbus_write
+    )
+    from .modbus_master_connection import ModbusConnectionManager
+except ImportError:
+    # Fallback to absolute imports (when run standalone)
+    from modbus_master_types import ModbusDeviceConfig
+    from modbus_master_utils import (
+        calculate_gcd_of_cycle_times,
+        get_batch_read_requests_from_io_points,
+        get_batch_write_requests_from_io_points,
+        get_modbus_registers_count_for_iec_size,
+        parse_modbus_offset
+    )
+    from modbus_master_memory import (
+        update_iec_buffer_from_modbus_data,
+        read_data_for_modbus_write
+    )
+    from modbus_master_connection import ModbusConnectionManager
 
 # Global variables for plugin lifecycle and configuration
 runtime_args = None
@@ -199,430 +71,25 @@ class ModbusSlaveDevice(threading.Thread):
         self.device_config = device_config
         self.sba = sba
         self._stop_event = threading.Event()
-        self.client: Optional[ModbusTcpClient] = None
+        self.connection_manager = ModbusConnectionManager(
+            device_config.host,
+            device_config.port,
+            device_config.timeout_ms
+        )
         self.name = f"ModbusSlave-{device_config.name}-{device_config.host}:{device_config.port}"
 
         # Calculate GCD of all I/O point cycle times for this device
         self.gcd_cycle_time_ms = calculate_gcd_of_cycle_times(device_config.io_points)
         print(f"[{self.name}] Calculated GCD cycle time: {self.gcd_cycle_time_ms}ms")
 
-        # Retry configuration - simple system
-        self.retry_delay_base = 2.0      # initial delay between attempts (seconds)
-        self.retry_delay_max = 30.0      # maximum delay between attempts (seconds)
-        self.retry_delay_current = self.retry_delay_base
-        self.is_connected = False
-
-    def _get_sba_access_details(self, iec_addr, is_write_op: bool = False) -> Optional[Dict[str, Any]]:
-        """
-        Maps IECAddress to SafeBufferAccess method parameters.
-        
-        Args:
-            iec_addr: IECAddress object
-            is_write_op: True if this is for a write operation (affects input/output buffer selection)
-        
-        Returns:
-            Dictionary with buffer access details or None if mapping fails
-        """
-        try:
-            area = iec_addr.area
-            size = iec_addr.size
-            
-            # Determine if this is a boolean operation
-            is_boolean = (size == "X")
-            
-            # Calculate buffer_idx based on size and index_bytes
-            if size == "X":  # Boolean - 1 bit
-                buffer_idx = iec_addr.index_bytes
-                bit_idx = iec_addr.bit
-                element_size_bytes = 1  # Bit operations work on byte boundaries
-            elif size == "B":  # Byte - 8 bits
-                buffer_idx = iec_addr.index_bytes
-                bit_idx = None
-                element_size_bytes = 1
-            elif size == "W":  # Word - 16 bits
-                buffer_idx = iec_addr.index_bytes // 2
-                bit_idx = None
-                element_size_bytes = 2
-            elif size == "D":  # Double word - 32 bits
-                buffer_idx = iec_addr.index_bytes // 4
-                bit_idx = None
-                element_size_bytes = 4
-            elif size == "L":  # Long word - 64 bits
-                buffer_idx = iec_addr.index_bytes // 8
-                bit_idx = None
-                element_size_bytes = 8
-            else:
-                print(f"[{self.name}]  Unsupported IEC size: {size}")
-                return None
-            
-            # Determine buffer type string based on area, size, and operation direction
-            if is_boolean:  # Size == "X"
-                if area == "I":
-                    buffer_type_str = "bool_input"
-                elif area == "Q":
-                    buffer_type_str = "bool_output"
-                elif area == "M":
-                    print(f"[{self.name}]  Memory area 'M' not supported for boolean operations")
-                    return None
-                else:
-                    print(f"[{self.name}]  Unknown area for boolean: {area}")
-                    return None
-            else:  # Non-boolean (B, W, D, L)
-                if area == "M":  # Memory area
-                    if size == "B":
-                        buffer_type_str = "byte_memory"  # Memory area uses memory buffer types
-                    elif size == "W":
-                        buffer_type_str = "int_memory"
-                    elif size == "D":
-                        buffer_type_str = "dint_memory"
-                    elif size == "L":
-                        buffer_type_str = "lint_memory"
-                    else:
-                        print(f"[{self.name}]  Unsupported memory size: {size}")
-                        return None
-                elif area == "I":  # Input area
-                    if size == "B":
-                        buffer_type_str = "byte_input"
-                    elif size == "W":
-                        buffer_type_str = "int_input"
-                    elif size == "D":
-                        buffer_type_str = "dint_input"
-                    elif size == "L":
-                        buffer_type_str = "lint_input"
-                    else:
-                        print(f"[{self.name}]  Unsupported input size: {size}")
-                        return None
-                elif area == "Q":  # Output area
-                    if size == "B":
-                        buffer_type_str = "byte_output"
-                    elif size == "W":
-                        buffer_type_str = "int_output"
-                    elif size == "D":
-                        buffer_type_str = "dint_output"
-                    elif size == "L":
-                        buffer_type_str = "lint_output"
-                    else:
-                        print(f"[{self.name}]  Unsupported output size: {size}")
-                        return None
-                else:
-                    print(f"[{self.name}]  Unknown area: {area}")
-                    return None
-            
-            return {
-                "buffer_type_str": buffer_type_str,
-                "buffer_idx": buffer_idx,
-                "bit_idx": bit_idx,
-                "element_size_bytes": element_size_bytes,
-                "is_boolean": is_boolean
-            }
-            
-        except Exception as e:
-            print(f"[{self.name}] (FAIL) Error in _get_sba_access_details: {e}")
-            return None
-
-    def _connect_with_retry(self) -> bool:
-        """
-        Attempts to connect to Modbus device with infinite retry.
-        
-        Returns:
-            True if connected successfully, False if thread was interrupted
-        """
-        host = self.device_config.host
-        port = self.device_config.port
-        timeout = self.device_config.timeout_ms / 1000.0
-        
-        retry_count = 0
-        
-        while not self._stop_event.is_set():
-            try:
-                # Create new client if necessary
-                if self.client is None or not self.client.connected:
-                    if self.client:
-                        try:
-                            self.client.close()
-                        except:
-                            pass
-                    self.client = ModbusTcpClient(host=host, port=port, timeout=timeout)
-                
-                # Attempt to connect
-                if self.client.connect():
-                    print(f"[{self.name}] (PASS) Connected to {host}:{port} (attempt {retry_count + 1})")
-                    self.is_connected = True
-                    self.retry_delay_current = self.retry_delay_base  # Reset delay
-                    return True
-                
-            except Exception as e:
-                print(f"[{self.name}] (FAIL) Connection attempt {retry_count + 1} failed: {e}")
-            
-            # Increment counter and calculate delay
-            retry_count += 1
-            
-            # Attempt logging
-            if retry_count == 1:
-                print(f"[{self.name}]  Failed to connect to {host}:{port}, starting retry attempts...")
-            elif retry_count % 10 == 0:  # Log every 10 attempts
-                print(f"[{self.name}]  Connection attempt {retry_count} failed, continuing retries...")
-            
-            # Wait with increasing delay (limited exponential backoff)
-            delay = min(self.retry_delay_current, self.retry_delay_max)
-            
-            # Sleep in small increments to allow quick stop
-            sleep_increments = int(delay * 10)  # 0.1s increments
-            for _ in range(sleep_increments):
-                if self._stop_event.is_set():
-                    return False
-                time.sleep(0.1)
-            
-            # Increase delay for next attempt (maximum of retry_delay_max)
-            self.retry_delay_current = min(self.retry_delay_current * 1.5, self.retry_delay_max)
-        
-        return False
-
     def _ensure_connection(self) -> bool:
         """
         Ensures there is a valid connection, reconnecting if necessary.
-        
+
         Returns:
             True if connection is available, False if thread was interrupted
         """
-        # Check if already connected
-        if self.client and self.client.connected:
-            return True
-        
-        # Mark as disconnected
-        self.is_connected = False
-        
-        # Try to reconnect
-        return self._connect_with_retry()
-
-    def _update_iec_buffer_from_modbus_data(self, iec_addr, modbus_data: list, length: int):
-        """
-        Updates IEC buffers with data read from Modbus.
-        Assumes mutex is already acquired.
-        
-        Args:
-            iec_addr: IECAddress object
-            modbus_data: List of values from Modbus (booleans for coils/inputs, integers for registers)
-            length: Number of IEC elements to write
-        """
-        try:
-            details = self._get_sba_access_details(iec_addr, is_write_op=True)
-            if not details:
-                print(f"[{self.name}] (FAIL) Failed to get SBA access details for {iec_addr}")
-                return
-            
-            buffer_type = details["buffer_type_str"]
-            base_buffer_idx = details["buffer_idx"]
-            base_bit_idx = details["bit_idx"]
-            is_boolean = details["is_boolean"]
-            iec_size = iec_addr.size
-            
-            # Write data elements to consecutive buffer locations
-            for i in range(length):
-                if is_boolean:
-                    # For boolean operations, handle bit indexing
-                    if i >= len(modbus_data):
-                        break  # No more data available
-                    
-                    current_data = modbus_data[i]
-                    
-                    if base_bit_idx is not None:
-                        # Calculate the actual bit position for this element
-                        current_bit_idx = base_bit_idx + i
-                        current_buffer_idx = base_buffer_idx + (current_bit_idx // 8)
-                        actual_bit_idx = current_bit_idx % 8
-                    else:
-                        current_buffer_idx = base_buffer_idx
-                        actual_bit_idx = i
-                    
-                    # Write boolean value
-                    if buffer_type == "bool_input":
-                        success, msg = self.sba.write_bool_input(current_buffer_idx, actual_bit_idx, 
-                                                                current_data, thread_safe=False)
-                    elif buffer_type == "bool_output":
-                        success, msg = self.sba.write_bool_output(current_buffer_idx, actual_bit_idx, 
-                                                                current_data, thread_safe=False)
-                    else:
-                        print(f"[{self.name}]  Unexpected boolean buffer type: {buffer_type}")
-                        continue
-                    
-                    if not success:
-                        print(f"[{self.name}] (FAIL) Failed to write boolean at buffer {current_buffer_idx}, bit {actual_bit_idx}: {msg}")
-                
-                else:
-                    # For non-boolean operations, handle register conversion
-                    registers_per_element = get_modbus_registers_count_for_iec_size(iec_size)
-                    start_reg_idx = i * registers_per_element
-                    end_reg_idx = start_reg_idx + registers_per_element
-                    
-                    if end_reg_idx > len(modbus_data):
-                        break  # Not enough register data available
-                    
-                    # Extract the registers for this IEC element
-                    element_registers = modbus_data[start_reg_idx:end_reg_idx]
-                    
-                    # Convert Modbus registers to IEC value
-                    try:
-                        if iec_size in ["B", "W"]:
-                            # For B and W, direct conversion (no multi-register)
-                            current_data = convert_modbus_registers_to_iec_value(element_registers, iec_size)
-                        elif iec_size in ["D", "L"]:
-                            # For D and L, combine multiple registers
-                            current_data = convert_modbus_registers_to_iec_value(element_registers, iec_size, use_big_endian=False)
-                        else:
-                            print(f"[{self.name}]  Unsupported IEC size: {iec_size}")
-                            continue
-                    except ValueError as e:
-                        print(f"[{self.name}] (FAIL) Error converting registers to IEC value: {e}")
-                        continue
-                    
-                    current_buffer_idx = base_buffer_idx + i
-                    
-                    # Write the value using the appropriate method
-                    if buffer_type == "byte_input":
-                        success, msg = self.sba.write_byte_input(current_buffer_idx, current_data, thread_safe=False)
-                    elif buffer_type == "byte_output":
-                        success, msg = self.sba.write_byte_output(current_buffer_idx, current_data, thread_safe=False)
-                    elif buffer_type == "int_input":
-                        success, msg = self.sba.write_int_input(current_buffer_idx, current_data, thread_safe=False)
-                    elif buffer_type == "int_output":
-                        success, msg = self.sba.write_int_output(current_buffer_idx, current_data, thread_safe=False)
-                    elif buffer_type == "int_memory":
-                        success, msg = self.sba.write_int_memory(current_buffer_idx, current_data, thread_safe=False)
-                    elif buffer_type == "dint_input":
-                        success, msg = self.sba.write_dint_input(current_buffer_idx, current_data, thread_safe=False)
-                    elif buffer_type == "dint_output":
-                        success, msg = self.sba.write_dint_output(current_buffer_idx, current_data, thread_safe=False)
-                    elif buffer_type == "dint_memory":
-                        success, msg = self.sba.write_dint_memory(current_buffer_idx, current_data, thread_safe=False)
-                    elif buffer_type == "lint_input":
-                        success, msg = self.sba.write_lint_input(current_buffer_idx, current_data, thread_safe=False)
-                    elif buffer_type == "lint_output":
-                        success, msg = self.sba.write_lint_output(current_buffer_idx, current_data, thread_safe=False)
-                    elif buffer_type == "lint_memory":
-                        success, msg = self.sba.write_lint_memory(current_buffer_idx, current_data, thread_safe=False)
-                    else:
-                        print(f"[{self.name}]  Unknown buffer type: {buffer_type}")
-                        continue
-                    
-                    if not success:
-                        print(f"[{self.name}] (FAIL) Failed to write {buffer_type} at index {current_buffer_idx}: {msg}")
-                        
-        except Exception as e:
-            print(f"[{self.name}] (FAIL) Error updating IEC buffer: {e}")
-
-    def _read_data_for_modbus_write(self, iec_addr, length: int) -> Optional[list]:
-        """
-        Reads data from IEC buffers for Modbus write operations.
-        Assumes mutex is already acquired.
-        
-        Args:
-            iec_addr: IECAddress object
-            length: Number of IEC elements to read
-        
-        Returns:
-            List of values ready for Modbus write or None if failed
-        """
-        try:
-            details = self._get_sba_access_details(iec_addr, is_write_op=False)
-            if not details:
-                print(f"[{self.name}] (FAIL) Failed to get SBA access details for {iec_addr}")
-                return None
-            
-            buffer_type = details["buffer_type_str"]
-            base_buffer_idx = details["buffer_idx"]
-            base_bit_idx = details["bit_idx"]
-            is_boolean = details["is_boolean"]
-            iec_size = iec_addr.size
-            
-            values = []
-            
-            # Read data elements from consecutive buffer locations
-            for i in range(length):
-                if is_boolean:
-                    # For boolean operations, handle bit indexing
-                    if base_bit_idx is not None:
-                        current_bit_idx = base_bit_idx + i
-                        current_buffer_idx = base_buffer_idx + (current_bit_idx // 8)
-                        actual_bit_idx = current_bit_idx % 8
-                    else:
-                        current_buffer_idx = base_buffer_idx
-                        actual_bit_idx = i
-                    
-                    # Read boolean value
-                    if buffer_type == "bool_input":
-                        value, msg = self.sba.read_bool_input(current_buffer_idx, actual_bit_idx, thread_safe=False)
-                    elif buffer_type == "bool_output":
-                        value, msg = self.sba.read_bool_output(current_buffer_idx, actual_bit_idx, thread_safe=False)
-                    else:
-                        print(f"[{self.name}]  Unexpected boolean buffer type: {buffer_type}")
-                        return None
-                    
-                    if msg != "Success":
-                        print(f"[{self.name}] (FAIL) Failed to read boolean at buffer {current_buffer_idx}, bit {actual_bit_idx}: {msg}")
-                        return None
-                    
-                    values.append(value)
-                
-                else:
-                    # For non-boolean operations
-                    current_buffer_idx = base_buffer_idx + i
-                    
-                    # Read the value using the appropriate method
-                    if buffer_type == "byte_input":
-                        value, msg = self.sba.read_byte_input(current_buffer_idx, thread_safe=False)
-                    elif buffer_type == "byte_output":
-                        value, msg = self.sba.read_byte_output(current_buffer_idx, thread_safe=False)
-                    elif buffer_type == "int_input":
-                        value, msg = self.sba.read_int_input(current_buffer_idx, thread_safe=False)
-                    elif buffer_type == "int_output":
-                        value, msg = self.sba.read_int_output(current_buffer_idx, thread_safe=False)
-                    elif buffer_type == "int_memory":
-                        value, msg = self.sba.read_int_memory(current_buffer_idx, thread_safe=False)
-                    elif buffer_type == "dint_input":
-                        value, msg = self.sba.read_dint_input(current_buffer_idx, thread_safe=False)
-                    elif buffer_type == "dint_output":
-                        value, msg = self.sba.read_dint_output(current_buffer_idx, thread_safe=False)
-                    elif buffer_type == "dint_memory":
-                        value, msg = self.sba.read_dint_memory(current_buffer_idx, thread_safe=False)
-                    elif buffer_type == "lint_input":
-                        value, msg = self.sba.read_lint_input(current_buffer_idx, thread_safe=False)
-                    elif buffer_type == "lint_output":
-                        value, msg = self.sba.read_lint_output(current_buffer_idx, thread_safe=False)
-                    elif buffer_type == "lint_memory":
-                        value, msg = self.sba.read_lint_memory(current_buffer_idx, thread_safe=False)
-                    else:
-                        print(f"[{self.name}]  Unknown buffer type: {buffer_type}")
-                        return None
-                    
-                    if msg != "Success":
-                        print(f"[{self.name}] (FAIL) Failed to read {buffer_type} at index {current_buffer_idx}: {msg}")
-                        return None
-                    
-                    # Convert IEC value to Modbus registers
-                    try:
-                        if iec_size in ["B", "W"]:
-                            # For B and W, direct conversion (single register)
-                            element_registers = convert_iec_value_to_modbus_registers(value, iec_size)
-                        elif iec_size in ["D", "L"]:
-                            # For D and L, split into multiple registers
-                            element_registers = convert_iec_value_to_modbus_registers(value, iec_size, use_big_endian=False)
-                        else:
-                            print(f"[{self.name}]  Unsupported IEC size: {iec_size}")
-                            return None
-                        
-                        # Add all registers for this element to the output list
-                        values.extend(element_registers)
-                        
-                    except ValueError as e:
-                        print(f"[{self.name}] (FAIL) Error converting IEC value to registers: {e}")
-                        return None
-            
-            return values
-                        
-        except Exception as e:
-            print(f"[{self.name}] (FAIL) Error reading data for Modbus write: {e}")
-            return None
+        return self.connection_manager.ensure_connection(self._stop_event)
 
     def run(self):
         print(f"[{self.name}] Thread started.")
@@ -635,7 +102,7 @@ class ModbusSlaveDevice(threading.Thread):
             return
 
         # Connect with infinite retry
-        if not self._connect_with_retry():
+        if not self.connection_manager.connect_with_retry(self._stop_event):
             print(f"[{self.name}] Thread stopped before connection could be established.")
             return
 
@@ -668,23 +135,8 @@ class ModbusSlaveDevice(threading.Thread):
                         continue
 
                     try:
-                        # Convert offset string to integer
-                        if not isinstance(point.offset, str) or not point.offset.strip():
-                            raise ValueError(f"Offset must be a non-empty string, got: {point.offset!r} (type: {type(point.offset)})")
-
-                        # Try to convert to integer, handling decimal and hexadecimal formats
-                        offset_str = point.offset.strip()
-                        try:
-                            # Support both decimal (123) and hexadecimal (0x1234, 0X1234) formats
-                            if offset_str.lower().startswith('0x'):
-                                address = int(offset_str, 16)  # Hexadecimal
-                            else:
-                                address = int(offset_str, 10)  # Decimal
-                        except ValueError as conv_err:
-                            raise ValueError(f"Cannot convert offset '{offset_str}' to integer (supports decimal or 0x hex): {conv_err}")
-
-                        if address < 0:
-                            raise ValueError(f"Offset must be non-negative, got: {address}")
+                        # Parse Modbus offset
+                        address = parse_modbus_offset(point.offset)
 
                         # Calculate the correct number of Modbus registers/coils needed
                         if point.fc in [3, 4]:  # Register-based operations (FC 3,4)
@@ -696,13 +148,13 @@ class ModbusSlaveDevice(threading.Thread):
 
                         # Perform Modbus read based on function code
                         if point.fc == 1:  # Read Coils
-                            response = self.client.read_coils(address, count)
+                            response = self.connection_manager.client.read_coils(address, count)
                         elif point.fc == 2:  # Read Discrete Inputs
-                            response = self.client.read_discrete_inputs(address, count)
+                            response = self.connection_manager.client.read_discrete_inputs(address, count)
                         elif point.fc == 3:  # Read Holding Registers
-                            response = self.client.read_holding_registers(address, count)
+                            response = self.connection_manager.client.read_holding_registers(address, count)
                         elif point.fc == 4:  # Read Input Registers
-                            response = self.client.read_input_registers(address, count)
+                            response = self.connection_manager.client.read_input_registers(address, count)
                         else:
                             print(f"[{self.name}] Unsupported read FC: {point.fc}")
                             continue
@@ -711,12 +163,12 @@ class ModbusSlaveDevice(threading.Thread):
                         if isinstance(response, (ModbusIOException, ExceptionResponse)):
                             print(f"[{self.name}] (FAIL) Modbus read error (FC {point.fc}, addr {address}): {response}")
                             # Mark as disconnected to force reconnection on next cycle
-                            self.is_connected = False
+                            self.connection_manager.is_connected = False
                             continue
                         elif response.isError():
                             print(f"[{self.name}] (FAIL) Modbus read failed (FC {point.fc}, addr {address}): {response}")
                             # Mark as disconnected to force reconnection on next cycle
-                            self.is_connected = False
+                            self.connection_manager.is_connected = False
                             continue
 
                         # Extract data from response
@@ -733,11 +185,11 @@ class ModbusSlaveDevice(threading.Thread):
                     except ConnectionException as ce:
                         print(f"[{self.name}] (FAIL) Connection error reading FC {point.fc}, offset {point.offset}: {ce}")
                         # Mark as disconnected to force reconnection
-                        self.is_connected = False
+                        self.connection_manager.is_connected = False
                     except Exception as e:
                         print(f"[{self.name}] (FAIL) Error reading FC {point.fc}, offset {point.offset}: {e}")
                         # For other errors also mark disconnected as precaution
-                        self.is_connected = False
+                        self.connection_manager.is_connected = False
 
                 # Batch update IEC buffers with single mutex acquisition
                 if read_results_to_update:
@@ -745,7 +197,7 @@ class ModbusSlaveDevice(threading.Thread):
                     if lock_acquired:
                         try:
                             for iec_addr, modbus_data, length in read_results_to_update:
-                                self._update_iec_buffer_from_modbus_data(iec_addr, modbus_data, length)
+                                update_iec_buffer_from_modbus_data(self.sba, iec_addr, modbus_data, length)
                         finally:
                             self.sba.release_mutex()
                     else:
@@ -766,23 +218,8 @@ class ModbusSlaveDevice(threading.Thread):
                         continue
 
                     try:
-                        # Convert offset string to integer
-                        if not isinstance(point.offset, str) or not point.offset.strip():
-                            raise ValueError(f"Offset must be a non-empty string, got: {point.offset!r} (type: {type(point.offset)})")
-
-                        # Try to convert to integer, handling decimal and hexadecimal formats
-                        offset_str = point.offset.strip()
-                        try:
-                            # Support both decimal (123) and hexadecimal (0x1234, 0X1234) formats
-                            if offset_str.lower().startswith('0x'):
-                                address = int(offset_str, 16)  # Hexadecimal
-                            else:
-                                address = int(offset_str, 10)  # Decimal
-                        except ValueError as conv_err:
-                            raise ValueError(f"Cannot convert offset '{offset_str}' to integer (supports decimal or 0x hex): {conv_err}")
-
-                        if address < 0:
-                            raise ValueError(f"Offset must be non-negative, got: {address}")
+                        # Parse Modbus offset
+                        address = parse_modbus_offset(point.offset)
 
                         # Read data from IEC buffers (with mutex)
                         lock_acquired, lock_msg = self.sba.acquire_mutex()
@@ -791,7 +228,7 @@ class ModbusSlaveDevice(threading.Thread):
                             continue
 
                         try:
-                            values_to_write = self._read_data_for_modbus_write(point.iec_location, point.length)
+                            values_to_write = read_data_for_modbus_write(self.sba, point.iec_location, point.length)
                         finally:
                             self.sba.release_mutex()
 
@@ -802,20 +239,20 @@ class ModbusSlaveDevice(threading.Thread):
                         # Perform Modbus write operation
                         if point.fc == 5:  # Write Single Coil
                             if len(values_to_write) > 0:
-                                response = self.client.write_coil(address, values_to_write[0])
+                                response = self.connection_manager.client.write_coil(address, values_to_write[0])
                             else:
                                 print(f"[{self.name}] (FAIL) No data to write for FC 5, offset {address}")
                                 continue
                         elif point.fc == 6:  # Write Single Register
                             if len(values_to_write) > 0:
-                                response = self.client.write_register(address, values_to_write[0])
+                                response = self.connection_manager.client.write_register(address, values_to_write[0])
                             else:
                                 print(f"[{self.name}] (FAIL) No data to write for FC 6, offset {address}")
                                 continue
                         elif point.fc == 15:  # Write Multiple Coils
-                            response = self.client.write_coils(address, values_to_write)
+                            response = self.connection_manager.client.write_coils(address, values_to_write)
                         elif point.fc == 16:  # Write Multiple Registers
-                            response = self.client.write_registers(address, values_to_write)
+                            response = self.connection_manager.client.write_registers(address, values_to_write)
                         else:
                             print(f"[{self.name}] Unsupported write FC: {point.fc}")
                             continue
@@ -824,22 +261,22 @@ class ModbusSlaveDevice(threading.Thread):
                         if isinstance(response, (ModbusIOException, ExceptionResponse)):
                             print(f"[{self.name}] (FAIL) Modbus write error (FC {point.fc}, addr {address}): {response}")
                             # Mark as disconnected to force reconnection on next cycle
-                            self.is_connected = False
+                            self.connection_manager.is_connected = False
                         elif response.isError():
                             print(f"[{self.name}] (FAIL) Modbus write failed (FC {point.fc}, addr {address}): {response}")
                             # Mark as disconnected to force reconnection on next cycle
-                            self.is_connected = False
+                            self.connection_manager.is_connected = False
 
                     except ValueError as ve:
                         print(f"[{self.name}] (FAIL) Invalid offset '{point.offset}' for FC {point.fc}: {ve}")
                     except ConnectionException as ce:
                         print(f"[{self.name}] (FAIL) Connection error writing FC {point.fc}, offset {point.offset}: {ce}")
                         # Mark as disconnected to force reconnection
-                        self.is_connected = False
+                        self.connection_manager.is_connected = False
                     except Exception as e:
                         print(f"[{self.name}] (FAIL) Error writing FC {point.fc}, offset {point.offset}: {e}")
                         # For other errors also mark disconnected as precaution
-                        self.is_connected = False
+                        self.connection_manager.is_connected = False
 
                 # 3. CYCLE TIMING - Sleep for GCD cycle time
                 cycle_elapsed = time.monotonic() - cycle_start_time
@@ -860,13 +297,12 @@ class ModbusSlaveDevice(threading.Thread):
         except ConnectionException as ce:
             print(f"[{self.name}] (FAIL) Connection failed: {ce}")
             # Try to reconnect
-            self.is_connected = False
+            self.connection_manager.is_connected = False
         except Exception as e:
             print(f"[{self.name}] (FAIL) Unexpected error in thread: {e}")
             traceback.print_exc()
         finally:
-            if self.client and self.client.connected:
-                self.client.close()
+            self.connection_manager.disconnect()
             print(f"[{self.name}] Thread finished and connection closed.")
 
     def stop(self):
@@ -917,7 +353,6 @@ def init(args_capsule):
         
     except Exception as e:
         print(f"(FAIL) Error during initialization: {e}")
-        import traceback
         traceback.print_exc()
         return False
 
@@ -954,7 +389,6 @@ def start_loop():
             
     except Exception as e:
         print(f"(FAIL) Error starting main loop: {e}")
-        import traceback
         traceback.print_exc()
         return False
 
@@ -999,7 +433,6 @@ def stop_loop():
         
     except Exception as e:
         print(f"(FAIL) Error stopping main loop: {e}")
-        import traceback
         traceback.print_exc()
         return False
